+++ front-end/src/stores/Effects.ts
@@ -1,0 +1,16 @@
+import { AtomEffect } from 'recoil';
+
+export const localStorageEffect: (key: string) => AtomEffect<any> =
+  (key: string) =>
+  ({ setSelf, onSet }) => {
+    const savedValue = localStorage.getItem(key);
+    if (savedValue != null) {
+      setSelf(JSON.parse(savedValue));
+    }
+
+    onSet((newValue, _, isReset) => {
+      isReset
+        ? localStorage.removeItem(key)
+        : localStorage.setItem(key, JSON.stringify(newValue));
+    });
+  };
